backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
PyMuPDF
sentence-transformers
faiss-cpu
numpy
